Update README.md
The purpose of this book is to illustrate techniques for conducting Rasch measurement theory [@Rasch_ori] analyses using existing R packages. The book includes some background information about Rasch models, and the primary objective is to demonstrate how to apply the models to data using R packages and how to interpret the results.

The primary audience for this book is graduate students or professionals who are familiar with Rasch measurement theory at a basic level, and who want to use the R software program to conduct their Rasch analyses. We provide a brief overview of several key features of Rasch measurement theory in this chapter, and we provide descriptions of basic characteristics of the models and analytic techniques in each of the following chapters. Accordingly, we encourage readers who are new to Rasch measurement theory to use this book as a supplement to other excellent introductory texts on the subject that include a detailed theoretical and statistical introduction to Rasch measurement.